requirements.txt
# Person VLM Requirements
# =======================

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0  # For pretrained vision models

# Transformers and tokenization
transformers>=4.35.0
tokenizers>=0.15.0

# Data handling
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0

# Training utilities
accelerate>=0.24.0
tqdm>=4.65.0
wandb>=0.16.0  # Optional: experiment tracking

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Inference and deployment (optional)
onnx>=1.15.0
onnxruntime>=1.16.0

# Development (optional)
pytest>=7.4.0
black>=23.0.0
